pushing code
unable to continue as of now. Constant migraines. Updated some code and have homepage working. Nothing else is working. Must re-read or find notes to see what to do. Will turn in project unfinished but will work on it throughout the week as it looks like project is being graded